+++ bika/lims/interfaces/__init__.py
@@ -908,3 +908,18 @@ class IWorkflowActionUIDsAdapter(IWorkflowActionAdapter):
     """Marker for adapters in charge of processing workflow action requests
     from views, but meant for redirection only
     """
+
+
+class IVerified(Interface):
+    """Marker interface for verified objects
+    """
+
+
+class ISubmitted(Interface):
+    """Marker interface for submitted objects
+    """
+
+
+class IReceived(Interface):
+    """Marker interface for received objects
+    """