+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
 
 **Changed**
 
+- #1330 Make guards to not rely on review history
 
 **Removed**